music_api/auth.py
from django.conf import settings
from rest_framework.permissions import BasePermission


class HasAPIKey(BasePermission):
    message = "Invalid or missing API key"

    def has_permission(self, request, view):
        api_key = request.headers.get("X-API-Key")
        return api_key == settings.API_KEY
